requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
pydantic==1.10.11
gunicorn==21.2.0
